Switch to Lambda Svelte adapter
Switch Svelte config to use @yarbsemaj/adapter-lambda. With this change,
pin the version of @sveltejs/kit to 1.0.0-next.201. This appears to be
the most recent version that is compatible with the most recent version
of the @yarbsemaj/adapter-lambda. Subsequent version break the build.

For more details, see the project page:

https://github.com/yarbsemaj/sveltekit-adapter-lambda